@@ -22,7 +22,8 @@
     "build": "npm run clean && npm run build-es && npm run build-cjs",
     "build-es": "tsc --noEmit false --module es2015 --target es2015 --outDir ./es -p tsconfig.json",
     "build-cjs": "tsc --noEmit false --module commonjs --target es5 --outDir ./lib -p tsconfig.json",
-    "build:examples": "cross-env vite build",
+    "build:example:legacy": "cross-env vite build -c vite.config-example-legacy.ts",
+    "build:example": "cross-env vite build -c vite.config-example.ts",
     "clean": "rimraf es lib build",
     "prepublish": "cross-env NODE_ENV=production npm run build",
     "lint": "eslint vite.config.js src examples --ext ts,tsx,js",
@@ -44,8 +45,8 @@
   ],
   "dependencies": {},
   "devDependencies": {
-    "@testing-library/react": "^12.1.2",
-    "@types/react": "^17.0.38",
+    "@testing-library/react": "^13.4.0",
+    "@types/react": "^18.0.0",
     "@typescript-eslint/parser": "^5.10.1",
     "@vitejs/plugin-react": "^1.1.4",
     "cross-env": "^5.0.1",
@@ -53,7 +54,7 @@
     "eslint-plugin-prettier": "^2.1.2",
     "jsdom": "^19.0.0",
     "prettier": "^1.5.3",
-    "react": "^16.8.0",
+    "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "rimraf": "^2.5.0",
     "typescript": "^4.5.5",
